Tolerate Map and Set in ddp-server data structures.
PR #7 partially accommodated the changes introduced by
https://github.com/meteor/meteor/pull/10053, but this commit seeks to
address the rest of the relevant changes, which mostly involve converting
objects to Map or Set.

Although these changes to ddp-server predated Meteor 1.8, the ddp-server
version was not bumped in Meteor 1.8 or 1.8.0.2, so Meteor 1.8.1 will be
the first time ddp-server is forced to update.